+++ packages/core/src/types/index.ts
@@ -62,6 +62,7 @@ export type ReversedDidList = Record<
 export enum NetworkName {
   mainnet = 'mainnet',
   testnet = 'testnet',
+  staging = 'staging',
   local = 'local',
 }